docs: enrich google consent screen docu
