+++ pkg/plugin/main.go
@@ -12,21 +12,29 @@ import (
 )
 
 //nolint:gochecknoglobals
+
 //goland:noinspection GoUnusedGlobalVariable
+
 var (
 	version = "dev"
-	commit  = "none"
-	date    = "unknown"
+
+	commit = "none"
+
+	date = "unknown"
 )
 
 type PluginHandle struct {
 	logger *slog.Logger
-	conf   config.Config
+
+	conf config.Config
+
 	server *httpserver.Server
 }
 
 func main() {
 	// This function is here to satisfy Go's requirement of having a main function.
+
 	// The main functionality is implemented in the openvpn_plugin_open_v3 function,
+
 	// which will be called from C.
 }
